Fixed 620 compile warnings

+++ ai/ai_monster.c
@@ -164,7 +164,7 @@ void AI_MonsterWorldEffects(edict_t* ent)
 			{	// suffocate!
 				if (ent->pain_debounce_time < level.time)
 				{
-					dmg = 2 + 2 * floor(level.time - ent->air_finished);
+					dmg = 2 + 2 * floorf(level.time - ent->air_finished);
 					if (dmg > 15)
 						dmg = 15;
 					Player_Damage(ent, world, world, vec3_origin, ent->s.origin, vec3_origin, dmg, 0, DAMAGE_NO_ARMOR, MOD_WATER);
@@ -188,7 +188,7 @@ void AI_MonsterWorldEffects(edict_t* ent)
 	{
 		if (ent->damage_debounce_time < level.time)
 		{
-			ent->damage_debounce_time = level.time + 0.2;
+			ent->damage_debounce_time = level.time + 0.2f;
 			Player_Damage(ent, world, world, vec3_origin, ent->s.origin, vec3_origin, 10 * ent->waterlevel, 0, 0, MOD_LAVA);
 		}
 	}
