+++ physics/physics_base.c
@@ -851,7 +851,7 @@ void SV_Physics_Step(edict_t* ent)
 	// friction for flying monsters that have been given vertical velocity
 	if ((ent->flags & FL_FLY) && (ent->velocity[2] != 0))
 	{
-		speed = fabs(ent->velocity[2]);
+		speed = fabsf(ent->velocity[2]);
 		control = speed < sv_stopspeed->value ? sv_stopspeed->value : speed;
 		friction = sv_friction->value / 3;
 		newspeed = speed - (FRAMETIME * control * friction);
@@ -864,7 +864,7 @@ void SV_Physics_Step(edict_t* ent)
 	// friction for flying monsters that have been given vertical velocity
 	if ((ent->flags & FL_SWIM) && (ent->velocity[2] != 0))
 	{
-		speed = fabs(ent->velocity[2]);
+		speed = fabsf(ent->velocity[2]);
 		control = speed < sv_stopspeed->value ? sv_stopspeed->value : speed;
 		newspeed = speed - (FRAMETIME * control * sv_waterfriction->value * ent->waterlevel);
 		if (newspeed < 0)
@@ -882,7 +882,7 @@ void SV_Physics_Step(edict_t* ent)
 			if (!(ent->health <= 0.0 && !M_CheckBottom(ent)))
 			{
 				vel = ent->velocity;
-				speed = sqrt(vel[0] * vel[0] + vel[1] * vel[1]);
+				speed = sqrtf(vel[0] * vel[0] + vel[1] * vel[1]);
 
 				if (speed)
 				{
